refactor: info service mock
