+++ apps/io-session-manager-internal/src/services/info.ts
@@ -1,6 +1,5 @@
 import * as RTE from "fp-ts/lib/ReaderTaskEither";
 import * as TE from "fp-ts/lib/TaskEither";
-import { pipe } from "fp-ts/lib/function";
 import { PackageUtils } from "../utils/package";
 import { ApplicationInfo } from "../generated/definitions/internal/ApplicationInfo";
